refactor: extract onboarding endpoints into feature-module package (#826)
Migrates /api/onboarding/* (status, interests, recommendations, profile,
source-recommendations) out of main.py into news_dashboard/onboarding/
(router/service/models), following the pattern established by the quizzes
reference module (#825). Behavior-preserving: no route paths or auth
semantics changed.
